Логика колоды карт вынесена в отдельный класс
