+++ src/main/java/org/example/controler/RideTheBus.java
@@ -1,114 +1,49 @@
 package org.example.controler;
 
+import org.example.game.Card;
+import org.example.game.Deck;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 
-import java.util.ArrayList;
-import java.util.List;
+
 import java.util.Random;
 
 /**
  * Класс Реализующий игру в Ride The Bus
  */
 public class RideTheBus {
-    private String[] values;
-    private String[] suits;
-    private Card[] table;
+    private Deck deck;
     private String specialCard;
-    private List<Card> deck;
-    private Random random;
     private final KeyboardFactory keyboardFactory;
-    private int roundNumber;
     private String chatId;
     private boolean isGameOver;
     private boolean isProcessing;
 
     public RideTheBus() {
         keyboardFactory = new KeyboardFactory();
-        values = new String[]{"2", "3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K", "A"};
-        suits = new String[]{"♠️", "♥️", "♦️", "♣️"};
         specialCard = "\uD83C\uDCCF";
-        random = new Random();
-        roundNumber = 1;
-        table = new Card[4];
+
         isGameOver = false;
         isProcessing = false;
-        initializeDeck();
+        deck = new Deck(4);
 
     }
 
-
-    /**
-     * Инициализация колоды
-     */
-    private void initializeDeck() {
-        deck = new ArrayList<>();
-        for (String suit : suits) {
-            for (String value : values) {
-                deck.add(new Card(value, suit));
-            }
-        }
-        shuffleDeck();
-    }
-
-    /**
-     *Используем алгоритм Фишера-Йейтса (Knuth shuffle) для перемешивания колоды
-     */
-    private void shuffleDeck() {
-        for (int i = deck.size() - 1; i > 0; i--) {
-            int randomIndex = random.nextInt(i + 1);
-            Card temp = deck.get(i);
-            deck.set(i, deck.get(randomIndex));
-            deck.set(randomIndex, temp);
-        }
-    }
-    /**
-     *Раздача карты
-     */
-    private Card dealCard() {
-        if (roundNumber == 1 && deck.size() < 52) {
-            initializeDeck();
-        }
-
-        return deck.remove(deck.size() - 1);
-    }
-    /**
-     *Добавление карты на стол
-     */
-    private void addToTable(Card card) {
-        for (int i = 0; i < 4; i++) {
-            if (table[i] == null) {
-                table[i] = card;
-                break;
-            }
-        }
-    }
-    /**
-     *Получение карт на столе как строку, а не как массив Card
-     */
-    private String getTableAsString(){
-        String tableAsString = "";
-        for (Card tableCard : table){
-            if (tableCard == null) break;
-            tableAsString += tableCard.getCard();
-            }
-        return tableAsString;
-    }
     /**
      *Проверка на победу
      */
     private boolean checkWin(String message, Card card) {
-        switch (roundNumber) {
+        switch (deck.roundNumber) {
             case 1:
                 boolean isRed = card.getSuit().equals("♥️") || card.getSuit().equals("♦️");
                 boolean isBlack = card.getSuit().equals("♠️") || card.getSuit().equals("♣️");
                 return (message.equals("red") && isRed) || (message.equals("black") && isBlack);
             case 2:
-                int valeCard1 = table[0].getValue();
+                int valeCard1 = deck.table[0].getValue();
                 int valueCurrent = card.getValue();
                 return ((valueCurrent >= valeCard1 && message.equals("higher")) || (valueCurrent <= valeCard1 && message.equals("lower")));
             case 3:
-                int valueCard1 = table[0].getValue();
-                int valueCard2 = table[1].getValue();
+                int valueCard1 = deck.table[0].getValue();
+                int valueCard2 = deck.table[1].getValue();
                 int currentValue = card.getValue();
                 int min = Math.min(valueCard1, valueCard2);
                 int max = Math.max(valueCard1, valueCard2);
@@ -140,10 +75,10 @@ public class RideTheBus {
      *Сброс состояния игры
      */
     private void resetGame() {
-        roundNumber = 1;
-        table = new Card[4];
-        initializeDeck();
-        random = new Random();
+        deck.roundNumber = 1;
+        deck = new Deck(4);
+        deck.initializeDeck();
+
         isGameOver = false;
         isProcessing = false; // Сброс флага обработки
     }
@@ -159,7 +94,7 @@ public class RideTheBus {
     private void play(TelegramBot bot) {
         SendMessage message = null;
 
-        switch (roundNumber) {
+        switch (deck.roundNumber) {
             case 1:
                 message = keyboardFactory.KeyboardFirstRound(chatId);
                 break;
@@ -183,7 +118,7 @@ public class RideTheBus {
         }
 
         if (message != null) {
-            message.setText(message.getText() + "\n" + getTableAsString() + " " + specialCard);
+            message.setText(message.getText() + "\n" + deck.getTableAsString() + " " + specialCard);
         }
         bot.send(message);
     }
@@ -195,20 +130,20 @@ public class RideTheBus {
             return;
         }
 
-        Card card = dealCard();
-        addToTable(card);
+        Card card = deck.dealCard();
+        deck.addToTable(card);
         boolean isWin = checkWin(callbackData, card);
         SendMessage winMessage = new SendMessage();
         winMessage.setChatId(chatId);
 
         if (isWin) {
-            winMessage.setText(getTableAsString() + "\nПоздравляем, вы выиграли!");
+            winMessage.setText(deck.getTableAsString() + "\nПоздравляем, вы выиграли!");
             bot.send(winMessage);
 
-            if (roundNumber == 4) {
+            if (deck.roundNumber == 4) {
                     handleGameOver(bot, true);
             } else {
-                roundNumber=roundNumber+1;
+                deck.roundNumber=deck.roundNumber+1;
                 play(bot);
             }
         } else {
@@ -224,9 +159,9 @@ public class RideTheBus {
         message.setChatId(chatId);
 
         if (isWinner) {
-            message.setText(getTableAsString() + "\nВы прошли все раунды! 🎉\nХотите выбрать другую игру?");
+            message.setText(deck.getTableAsString() + "\nВы прошли все раунды! 🎉\nХотите выбрать другую игру?");
         } else {
-            message.setText(getTableAsString() + "\nК сожалению, вы проиграли! Хотите сыграть снова?");
+            message.setText(deck.getTableAsString() + "\nК сожалению, вы проиграли! Хотите сыграть снова?");
         }
 
         message.setReplyMarkup(keyboardFactory.createGameSelectionKeyboard(chatId).getReplyMarkup());
