+++ src/main/java/org/example/game/Card.java
@@ -1,4 +1,4 @@
-package org.example.controler;
+package org.example.game;
 
 /**
  * Класс, реализующий представление карты
